refactored booking service and adapted tests accordingly
